refactor(landingPageTop) - Changing name function to english
